replaced post screen with post widget
