+++ app/lib/screens/home_screen.dart
@@ -4,7 +4,7 @@ import '../models/user.dart';
 import '../services/auth_service.dart';
 import '../utils/fade_route.dart';
 import 'login_screen.dart';
-import 'post_screen.dart';
+import '../components/post.dart';
 
 class HomeScreen extends StatelessWidget {
   final User user;
@@ -36,7 +36,7 @@ class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final screenHeight = MediaQuery.of(context).size.height;
-    final imageHeight = screenHeight * 0.7;
+    final postHeight = screenHeight * 0.7;
 
     return Scaffold(
       backgroundColor: const Color(0xFF1A1A1A),
@@ -56,78 +56,12 @@ class HomeScreen extends StatelessWidget {
       body: ListView.builder(
         itemCount: samplePosts.length,
         itemBuilder: (context, index) {
-          return GestureDetector(
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => PostScreen(
-                    imageUrl: samplePosts[index],
-                  ),
-                ),
-              );
-            },
-            child: SizedBox(
-              height: imageHeight,
-              child: Card(
-                margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-                clipBehavior: Clip.antiAlias,
-                color: const Color(0xFF242424),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Expanded(
-                      child: Center(
-                        child: Image.network(
-                          samplePosts[index],
-                          fit: BoxFit.contain,
-                          loadingBuilder: (context, child, loadingProgress) {
-                            if (loadingProgress == null) return child;
-                            return const Center(
-                              child: CircularProgressIndicator(
-                                valueColor: AlwaysStoppedAnimation<Color>(
-                                    Color(0xFF4CAF50)),
-                              ),
-                            );
-                          },
-                          errorBuilder: (context, error, stackTrace) {
-                            return const Center(
-                              child: Text(
-                                'Failed to load image',
-                                style: TextStyle(color: Colors.white70),
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.all(16.0),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            'Post ${index + 1}',
-                            style: const TextStyle(
-                              color: Colors.white70,
-                              fontSize: 18,
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                          const SizedBox(height: 8),
-                          Text(
-                            'Posted by ${user.username}',
-                            style: const TextStyle(
-                              color: Colors.grey,
-                              fontSize: 14,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
+          return SizedBox(
+            height: postHeight,
+            child: Post(
+              imageUrl: samplePosts[index],
+              username: user.username,
+              title: 'Post ${index + 1}',
             ),
           );
         },
