--- app/lib/screens/post_screen.dart
@@ -1,47 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PostScreen extends StatelessWidget {
-  final String imageUrl;
-
-  const PostScreen({
-    Key? key,
-    required this.imageUrl,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF1A1A1A),
-      appBar: AppBar(
-        backgroundColor: const Color(0xFF242424),
-        title: const Text(
-          'Post',
-          style: TextStyle(color: Colors.white70),
-        ),
-        iconTheme: const IconThemeData(color: Colors.white70),
-      ),
-      body: Center(
-        child: Image.network(
-          imageUrl,
-          fit: BoxFit.contain,
-          loadingBuilder: (context, child, loadingProgress) {
-            if (loadingProgress == null) return child;
-            return const Center(
-              child: CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation<Color>(Color(0xFF4CAF50)),
-              ),
-            );
-          },
-          errorBuilder: (context, error, stackTrace) {
-            return const Center(
-              child: Text(
-                'Failed to load image',
-                style: TextStyle(color: Colors.white70),
-              ),
-            );
-          },
-        ),
-      ),
-    );
-  }
-}
